requirements.txt
streamlit==1.38.0 
pandas==2.2.2
openai==1.45.0
langchain==0.3.0
chromadb==0.5.3
langchain-chroma==0.1.4
langchain-openai==0.2.0
python-dotenv==1.0.1
openpyxl==3.1.5
XlsxWriter==3.2.0
bs4==0.0.2
langchain-community==0.3.0
langchain-mongodb==0.2.0
vaderSentiment==3.3.2
matplotlib==3.9.2